refact(manager): factor out permissions list cell component
